Make the GraphQL grammar header public
